pyproject.toml
[project]
name = "digital-direction"
version = "0.1.0"
description = "Telecom document extraction pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "pdfplumber>=0.11.0",
    "docling>=2.0.0",
    "google-generativeai>=0.8.0",
    "anthropic>=0.40.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    "pandas>=2.2.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "alembic>=1.13.0",
    "extract-msg>=0.50.0",
    "langfuse>=2.0.0",
]

[tool.setuptools.packages.find]
include = ["backend*", "evals*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
